src/api/requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
python-multipart==0.0.6
requests==2.31.0
transformers==4.18.0
torch==2.0.1
Pillow==10.0.0
open-clip-torch==2.20.0
numpy==1.24.3
langchain==0.0.217
pymilvus==2.2.13
pandas==2.0.3
sentence_transformers==2.2.2
openai==0.27.8

